% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extract Coefficients from Lasso Model}
\usage{
extract_coefficients(cv_fit)
}
\arguments{
\item{cv_fit}{Cross-validated Lasso model object (from cv.glmnet).}
}
\value{
Sparse matrix of coefficients at lambda.min.
}
\description{
Extracts coefficients from a cross-validated lasso model at lambda.min.
}
\keyword{internal}
