% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{extract_grp_lasso_coefficients}
\alias{extract_grp_lasso_coefficients}
\title{Extract Coefficients from Group Lasso Model}
\usage{
extract_grp_lasso_coefficients(cv_fit, recovery_index)
}
\arguments{
\item{cv_fit}{Cross-validated Group Lasso model object (from cv.grpnet).}

\item{recovery_index}{Index vector to recover original coefficient order.}
}
\value{
Vector of coefficients for original and knockoff SNPs and their interaction terms.
}
\description{
Extracts coefficients from a cross-validated group lasso model and reorders
them to match the original predictor order.
}
\keyword{internal}
