% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{generate_knockoff_data}
\alias{generate_knockoff_data}
\title{Generate Knockoff Data for SNPs}
\usage{
generate_knockoff_data(genetic_variants)
}
\arguments{
\item{genetic_variants}{Matrix of genetic variants (SNPs) where rows are
samples and columns are SNPs.}
}
\value{
Knockoff genotype matrix with the same dimensions as the original
genotype matrix (n_samples x n_SNPs).
}
\description{
This function takes a genotype matrix and generates knockoff data using
second-order knockoffs with semidefinite programming (SDP) method.
}
