% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap of Selected SNPs}
\usage{
plot_heatmap(
  scaled_selection_matrix,
  genetic_variants,
  selected_snp_names,
  eur,
  save_path = "."
)
}
\arguments{
\item{scaled_selection_matrix}{Scaled selection matrix from the knockoff filter
(n_samples x n_variants).}

\item{genetic_variants}{Matrix of genetic variants (n_samples x n_variants).
Column names are used to match with selected_snp_names.}

\item{selected_snp_names}{Character vector of names of selected SNPs to plot.}

\item{eur}{Numeric vector representing EUR percentages to sort individuals
(length n_samples).}

\item{save_path}{The directory path where the plot will be saved
(default: current working directory).}
}
\value{
A ggplot object representing the heatmap.
}
\description{
Creates a heatmap of selected SNPs based on the scaled selection matrix,
sorted by EUR percentages for individuals.
}
