% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{prepare_feat_grps}
\alias{prepare_feat_grps}
\title{Prepare Groups Information for Group Lasso Model}
\usage{
prepare_feat_grps(num_snps, num_z, num_u, num_all)
}
\arguments{
\item{num_snps}{Number of SNPs (equal to ncol(genetic_variants)).}

\item{num_z}{Number of interacting variables (number of PCs).}

\item{num_u}{Number of unpenalized covariates (sum of additional vars and Z).}

\item{num_all}{Total number of predictors (first dimension of matrix X).}
}
\value{
List with components:
\item{new_order}{Vector of new column order for predictors.}
\item{recovery_index}{Index to recover original coefficient order.}
\item{grp_for_grplasso}{Group indices for group lasso fitting.}
}
\description{
Prepares group structure information for group lasso modeling.
}
\keyword{internal}
