% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{prepare_interaction_terms}
\alias{prepare_interaction_terms}
\title{Prepare Interaction Terms}
\usage{
prepare_interaction_terms(genetic_variants, Z)
}
\arguments{
\item{genetic_variants}{Matrix of original or knockoff genetic variants
(n_samples x n_variants).}

\item{Z}{Vector or matrix representing heterogeneity variable(s)
(e.g., EUR or 4 PCs) (n_samples x n_Z_variables).}
}
\value{
Matrix of interaction terms (n_samples x (n_variants * n_Z_variables)).
}
\description{
Creates interaction terms between genetic variants and heterogeneity variables.
}
\keyword{internal}
