% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{set_penalty_factors_grpLasso}
\alias{set_penalty_factors_grpLasso}
\title{Set Penalty Factors for Group Lasso}
\usage{
set_penalty_factors_grpLasso(num_snps, num_z, num_u)
}
\arguments{
\item{num_snps}{Number of SNPs (equal to ncol(genetic_variants)).}

\item{num_z}{Number of interacting variables (number of PCs).}

\item{num_u}{Number of unpenalized covariates (sum of additional vars and Z).}
}
\value{
Vector of penalty factors with length equal to total number of groups.
}
\description{
Creates penalty factors vector for group lasso regression. Uses square-root
of group size as the default penalty for groups (except unpenalized covariates).
}
\keyword{internal}
